fix: 🐛 merge conflicts and use app-table-row in tables
